Add WriteVTK extension module
