+++ Project.toml
@@ -18,13 +18,14 @@ Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
 PartitionedArrays = "5a9dfac6-5c52-46f7-8278-5e2210713be9"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
-WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
 [weakdeps]
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
 [extensions]
 GalerkinToolkitMakieExt = "Makie"
+GalerkinToolkitWriteVTKExt = "WriteVTK"
 
 [compat]
 Combinatorics = "1"
@@ -33,16 +34,17 @@ ForwardDiff = "0.10"
 Gmsh = "0.3"
 IterativeSolvers = "0.9"
 MPI = "0.20"
+Makie = "0.21"
+WriteVTK = "1.21"
 Metis = "1"
 PartitionedArrays = "0.5.4"
 StaticArrays = "1"
-WriteVTK = "1"
-Makie = "0.21"
 julia = "1"
 
 [extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "GLMakie"]
+test = ["Test", "GLMakie", "WriteVTK"]
